requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.1
python-multipart==0.0.6
aiohttp==3.8.6
websockets==11.0.3